khetwala-backend/requirements.txt
# Khetwala Backend Dependencies
# ══════════════════════════════════════════════════════════════════════════════

# ─── Core Framework ───────────────────────────────────────────────────────────
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0

# ─── HTTP Clients ─────────────────────────────────────────────────────────────
requests>=2.31.0,<3.0
httpx>=0.27.0,<1.0
aiohttp>=3.9.0,<4.0

# ─── Configuration & Environment ──────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0

# ─── Caching ──────────────────────────────────────────────────────────────────
cachetools>=5.3.0,<6.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.25,<3.0
alembic>=1.13.0,<2.0
# psycopg2-binary>=2.9.9,<3.0  # Uncomment if using PostgreSQL

# ─── ETL & Scheduling ────────────────────────────────────────────────────────
apscheduler>=3.10.4,<4.0

# ─── Machine Learning ─────────────────────────────────────────────────────
scikit-learn>=1.4.0,<2.0
numpy>=1.26.0,<3.0
xgboost>=2.0.0,<3.0

# ─── Production Server ────────────────────────────────────────────────────────
gunicorn>=21.0.0,<23.0

# ─── Logging & Monitoring ─────────────────────────────────────────────────────
structlog>=24.1.0,<25.0
python-json-logger>=2.0.0,<3.0

# ─── Security & Rate Limiting ─────────────────────────────────────────────────
slowapi>=0.1.9,<1.0

# ─── Authentication ───────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0
python-multipart>=0.0.7,<1.0

# ─── Development & Testing ────────────────────────────────────────────────────
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0
pytest-cov>=4.1.0,<5.0
